[Experimental] Instead of .scroll use .animate to adjust the margin-left of the slider
This seems to be a lot smoother; however it makes the size of the page shrink and grow, which may have side-effects.

An interesting side-effect is that when you hit the right-hand edge, it appears to "bounce off". It looks cool but only at high scroll speeds and only if you keep
going right; otherwise it just bounces off once and in some cases you can end up further left than you started, which is confusing.