'x' key on keyboard (intentionally no joystick equivalent) reverses the meaning of "left" and "right"
This year we flip the projected image because it's back-projected and (unlike last year) there's text on the screen for the Easter Egg. So left and right
need to be swapped, but I like being able to control that from a keyboard. It defaults to true though.

Debug window hidden, and the dbg() function is no-op when that's the case at startup, rather than checking every time.