Exit from space is now allowed by going down from the entry point.
This required some special handling to only action it when close to the bottom. Also normally we don't follow area changes if a direction was held down,
to avoid continuously going through levels, but that's the only way to exit from this level.

We also want to show the arrow in advance of the actual exit point to make it more obvious.