Try to fix Pi undersea performance
Moving around undersea on the Pi started off fine but once you'd gone a certain distance the scrolling slowed to a crawl.

Instead of using a wide div and moving things off screen, the div is now only a screen width but the script pretends it's larger for the purposes of
calculating centerX and the object positions. Objects are also hidden once they go out of view.

Also scaled the chain and anchor PNG to the size we actually use it, instead of dynamically doing so. The image is now about 1/3 the filesize.