Get rid of the "smooth scrolling" margin transitions, because they actually had the opposite effect!
In main, the margin-left transition caused a very visible content flash whenever images were rotated and centerX recalculated.
It was bad.
And in any case it didn't make the scrolling any less steppy on the Pi, which is caused by low frame rate and applies to animations too.
So removed the margin-top one as well.